requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cloudinary==1.44.1
cryptography==46.0.5
Django==5.2.12
django-cloudinary-storage==0.3.0
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
pycparser==3.0
PyMySQL==1.1.2
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
